+++ bin/monorepo-builder.php
@@ -3,7 +3,26 @@
 declare(strict_types=1);
 
 use Symplify\MonorepoBuilder\Config\MBConfig;
+use Symplify\MonorepoBuilder\Release\ReleaseWorker\PushTagReleaseWorker;
+use Symplify\MonorepoBuilder\Release\ReleaseWorker\SetCurrentMutualDependenciesReleaseWorker;
+use Symplify\MonorepoBuilder\Release\ReleaseWorker\SetNextMutualDependenciesReleaseWorker;
+use Symplify\MonorepoBuilder\Release\ReleaseWorker\UpdateBranchAliasReleaseWorker;
+use Symplify\MonorepoBuilder\Release\ReleaseWorker\UpdateReplaceReleaseWorker;
+use Config\MonorepoBuilder\AddTagToChangelogReleaseWorker;
+use Config\MonorepoBuilder\PushNextDevReleaseWorker;
+use Config\MonorepoBuilder\TagVersionReleaseWorker;
 
 return static function (MBConfig $mbConfig): void {
     $mbConfig->packageDirectories([dirname(__DIR__) . '/src']);
+    // release workers - in order to execute
+    $mbConfig->workers([
+        UpdateReplaceReleaseWorker::class,
+        SetCurrentMutualDependenciesReleaseWorker::class,
+        AddTagToChangelogReleaseWorker::class,
+        TagVersionReleaseWorker::class,
+        SetNextMutualDependenciesReleaseWorker::class,
+        UpdateBranchAliasReleaseWorker::class,
+        PushNextDevReleaseWorker::class,
+        PushTagReleaseWorker::class,
+    ]);
 };